tests: moved and removed superfluous tests
